requirements.txt
keyboard==0.13.5
numpy==1.26.4
Pillow==10.3.0
pyautogui==0.9.54
pytesseract==0.3.10
opencv-python
